refactor(pages/api): 전면 api 삭제 후 actions 수정

+++ frontend/src/actions/auth/auth.js
@@ -1,3 +1,5 @@
+import Cookies from 'js-cookie'
+import { API_URL } from '../../config/index';
 import { 
     REGISTER_SUCCESS,
     REGISTER_FAIL,
@@ -40,7 +42,7 @@ export const register = (registerData, callback) => async dispatch => {
     });
 
     try {
-        const res = await fetch('/api/user/save', {
+        const res = await fetch(`${API_URL}/api/user/save`, {
             method: 'POST',
             headers: {
                 'Accept': 'application/json',
@@ -49,25 +51,22 @@ export const register = (registerData, callback) => async dispatch => {
             body: body
         });
 
-        const data = await res.json();
+        const apiRes = await res.json();
 
         dispatch({
             type: REMOVE_AUTH_LOADING
         });
         
         if (res.status === 201) {
-            console.log("suc");
             dispatch({
                 type: REGISTER_SUCCESS
             });
-            if (callback) callback([true, data.success]);
+            if (callback) callback([true, apiRes.message]);
         } else {
-            console.log("fail");
             dispatch({
-                type: REGISTER_FAIL,
-                payload: data
+                type: REGISTER_FAIL
             });
-            if (callback) callback([false, data.error]);
+            if (callback) callback([false, apiRes.message]);
         }
     } catch(error) {
         console.log(error);
@@ -89,7 +88,7 @@ export const login = (username, password, callback) => async dispatch => {
     });
 
     try {
-        const res = await fetch('/api/user/login', {
+        const res = await fetch(`${API_URL}/api/user/login`, {
             method: 'POST',
             headers: {
                 'Accept': 'application/json',
@@ -97,26 +96,28 @@ export const login = (username, password, callback) => async dispatch => {
             },
             body: body
         });
+
+        const apiRes = await res.json();
         
-        console.log(res)
-        const data = await res.json();
-        console.log(data)
-        
+        console.log(apiRes)
+
         dispatch({
             type: REMOVE_AUTH_LOADING
         });
 
         if (res.status === 200) {
             dispatch({
-                type: LOGIN_SUCCESS
+                type: LOGIN_SUCCESS,
+                payload: apiRes.data
             });
-            // dispatch(load_user());
-            if (callback) callback([true, data.success]);
+            dispatch(load_user());
+            if (callback) callback([true, apiRes.message]);
+
         } else {
             dispatch({
                 type: LOGIN_FAIL
             });
-            if (callback) callback([false, data.error]);
+            if (callback) callback([false, apiRes.message]);
         }
     } catch(error) {
         console.log(error);
@@ -129,22 +130,9 @@ export const login = (username, password, callback) => async dispatch => {
 
 export const logout = () => async dispatch => {
     try {
-        const res = await fetch('/api/user/logout', {
-            method: 'POST',
-            headers: {
-                'Accept': 'application/json',
-            }
+        dispatch({
+            type: LOGOUT_SUCCESS
         });
-
-        if (res.status === 200){
-            dispatch({
-                type: LOGOUT_SUCCESS
-            });
-        } else {
-            dispatch({
-                type: LOGOUT_FAIL
-            });
-        }
     } catch(error){
         console.log(error);
         dispatch({
@@ -154,25 +142,35 @@ export const logout = () => async dispatch => {
 }
 
 export const load_user = () => async dispatch => {
+    const access = Cookies.get('access') ?? null;
+
+    if (access === null) {
+        console.log('access 토큰이 존재하지 않습니다')
+        return dispatch({
+            type: AUTHENTICATED_FAIL
+        });
+    }
+
     try {
-        const res = await fetch('/api/user/me',{
+        const res = await fetch(`${API_URL}/api/user/me`,{
             method: 'GET',
             headers: {
-                'Accept': 'application/json'
+                'Accept': 'application/json',
+                'Authorization' : `Bearer ${access}`
             }
         });
 
-        const data = await res.json();
+        const apiRes = await res.json();
 
         if (res.status === 200) {
             dispatch({
                 type: LOAD_USER_SUCCESS,
-                payload: data
+                payload: apiRes.data
             });
         } else {
             dispatch({
                 type: LOAD_USER_FAIL,
-                payload: data
+                payload: apiRes.data
             });
         }
 
@@ -186,11 +184,22 @@ export const load_user = () => async dispatch => {
 
 
 export const request_verify = () => async dispatch => {
+    const access = Cookies.get('access') ?? null;
+
+    if (access === null) {
+        console.log('access 토큰이 존재하지 않습니다');
+        return dispatch({
+            type: AUTHENTICATED_FAIL
+        });
+    }
+    
+
     try {
-        const res = await fetch('/api/user/token/verify',{
+        const res = await fetch(`${API_URL}/api/user/token/verify`,{
             method: 'GET',
             headers: {
                 'Accept' : 'application/json',
+                'Authorization' : `Bearer ${access}`
             }
         });
 
@@ -198,12 +207,10 @@ export const request_verify = () => async dispatch => {
             dispatch({
                 type: AUTHENTICATED_SUCCESS
             });
-            dispatch(load_user());
         } else {
             dispatch({
                 type: AUTHENTICATED_FAIL
             });
-            dispatch(request_verify());
         }
     } catch (error) {
         console.log(error);
@@ -214,25 +221,34 @@ export const request_verify = () => async dispatch => {
 }
 
 export const request_refresh = () => async dispatch => {
+    const refresh = Cookies.get('refresh') ?? null;
+
+    if (refresh === null) {
+        console.log('refresh 토큰이 존재하지 않습니다');
+        return dispatch({
+            type: AUTHENTICATED_FAIL
+        });
+    }
+
     try {
-        const res = await fetch('/api/user/token/refresh', {
+        const res = await fetch(`${API_URL}/api/user/token/refresh`, {
             method: 'GET',
             headers: {
                 'Accept': 'application/json',
+                'Authorization' : `Bearer ${refresh}`
             }
         });
 
-        const data = await res.json();
+        const apiRes = await res.json();
 
-        if(res.status === 200){
+        if (res.status === 200) {
             dispatch({
-                type: REFRESH_SUCCESS
+                type: REFRESH_SUCCESS,
+                payload: apiRes.data
             });
-            dispatch(request_verify());
-        } else{
+        } else {
             dispatch({
-                type: REFRESH_FAIL,
-                payload: data
+                type: REFRESH_FAIL
             });
         }
     } catch (error) {
@@ -249,7 +265,7 @@ export const check_username = (username, callback) => async dispatch => {
     });
 
     try {
-        const res = await fetch('/api/user/check/username', {
+        const res = await fetch(`${API_URL}/api/user/check/username`, {
             method: 'POST',
             headers: {
                 'Accept': 'application/json',
@@ -258,20 +274,20 @@ export const check_username = (username, callback) => async dispatch => {
             body: body
         });
 
-        const data = await res.json();
+        const apiRes = await res.json();
 
         if(res.status === 200){
             dispatch({
                 type: CHECK_USERNAME_SUCCESS,
-                payload: data
+                payload: apiRes.data
             });
-            if (callback) callback([true, data.success]);
+            if (callback) callback([true, apiRes.message]);
         } else{
             dispatch({
                 type: CHECK_USERNAME_FAIL,
-                payload: data
+                payload: apiRes.data
             });
-            if (callback) callback([false, data.error]);
+            if (callback) callback([false, apiRes.message]);
         }
     } catch (error) {
         console.log(error);
@@ -288,7 +304,7 @@ export const check_nickname = (nickname, callback) => async dispatch => {
     });
 
     try {
-        const res = await fetch('/api/user/check/nickname', {
+        const res = await fetch(`${API_URL}/api/user/check/nickname`, {
             method: 'POST',
             headers: {
                 'Accept': 'application/json',
@@ -297,20 +313,20 @@ export const check_nickname = (nickname, callback) => async dispatch => {
             body: body
         });
 
-        const data = await res.json();
+        const apiRes = await res.json();
 
         if(res.status === 200){
             dispatch({
                 type: CHECK_NICKNAME_SUCCESS,
-                payload: data
+                payload: apiRes.data
             });
-            if (callback) callback([true, data.success]);
+            if (callback) callback([true, apiRes.message]);
         } else{
             dispatch({
                 type: CHECK_NICKNAME_FAIL,
-                payload: data
+                payload: apiRes.data
             });
-            if (callback) callback([false, data.error]);
+            if (callback) callback([false, apiRes.message]);
         }
     } catch (error) {
         console.log(error);
@@ -322,33 +338,43 @@ export const check_nickname = (nickname, callback) => async dispatch => {
 }
 
 export const change_user = (nickname, major, image, student_id, callback) => async dispatch => {
+    const access = Cookies.get('access') ?? null;
+
+    if (access === null) {
+        console.log('access 토큰이 존재하지 않습니다');
+        return dispatch({
+            type: AUTHENTICATED_FAIL
+        });
+    }
+
     const body = JSON.stringify({
         nickname, major, image, student_id
     });
 
     try {
-        const res = await fetch('/api/user/me', {
+        const res = await fetch(`${API_URL}/api/user/me`, {
             method: 'PUT',
             headers: {
                 'Accept': 'application/json',
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
+                'Authorization' : `Bearer ${access}`
             },
             body: body
         });
 
-        const data = await res.json();
+        const apiRes = await res.json();
 
         if(res.status === 200){
             dispatch({
                 type: CHANGE_USER_SUCCESS
             });
-            if (callback) callback([true, data.success]);
+            if (callback) callback([true, apiRes.message]);
         } else{
             dispatch({
                 type: CHANGE_USER_FAIL,
-                payload: data
+                payload: apiRes.data
             });
-            if (callback) callback([false, data.error]);
+            if (callback) callback([false, apiRes.message]);
         }
     } catch (error) {
         console.log(error);
@@ -360,6 +386,15 @@ export const change_user = (nickname, major, image, student_id, callback) => asy
 }
 
 export const change_password = (password, new_password, new_re_password, callback) => async dispatch => {
+    const access = Cookies.get('access') ?? null;
+
+    if (access === null) {
+        console.log('access 토큰이 존재하지 않습니다');
+        return dispatch({
+            type: AUTHENTICATED_FAIL
+        });
+    }
+
     const body = JSON.stringify({
         password,
         new_password,
@@ -367,29 +402,30 @@ export const change_password = (password, new_password, new_re_password, callbac
     });
 
     try {
-        const res = await fetch('/api/user/password', {
+        const res = await fetch(`${API_URL}/api/user/password`, {
             method: 'PUT',
             headers: {
                 'Accept': 'application/json',
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
+                'Authorization' : `Bearer ${access}`
             },
             body: body
         });
 
-        const data = await res.json();
+        const apiRes = await res.json();
 
         if(res.status === 200){
             dispatch({
                 type: CHANGE_PASSWORD_SUCCESS,
-                payload: data
+                payload: apiRes.data
             });
-            if (callback) callback([true, data.success]);
+            if (callback) callback([true, apiRes.message]);
         } else{
             dispatch({
                 type: CHANGE_PASSWORD_FAIL,
-                payload: data
+                payload: apiRes.data
             });
-            if (callback) callback([false, data.error]);
+            if (callback) callback([false, apiRes.message]);
         }
     } catch (error) {
         console.log(error);
@@ -401,16 +437,26 @@ export const change_password = (password, new_password, new_re_password, callbac
 }
 
 export const change_toggle = (campus) => async dispatch => {
+    const access = Cookies.get('access') ?? null;
+
+    if (access === null) {
+        console.log('access 토큰이 존재하지 않습니다');
+        return dispatch({
+            type: AUTHENTICATED_FAIL
+        });
+    }
+
     const body = JSON.stringify({
         campus
     });
 
     try {
-        const res = await fetch('/api/user/toggle', {
+        const res = await fetch(`${API_URL}/api/user/toggle`, {
             method: 'PUT',
             headers: {
                 'Accept': 'application/json',
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
+                'Authorization' : `Bearer ${access}`
             },
             body: body
         });
@@ -433,26 +479,37 @@ export const change_toggle = (campus) => async dispatch => {
 }
 
 export const delete_user = (callback) => async dispatch => {
+    const access = Cookies.get('access') ?? null;
+
+    if (access === null) {
+        console.log('access 토큰이 존재하지 않습니다');
+        return dispatch({
+            type: AUTHENTICATED_FAIL
+        });
+    }
 
     try {
-        const res = await fetch('/api/user/me', {
-            method: 'DELETE'
+        const res = await fetch(`${API_URL}/api/user/me`, {
+            method: 'DELETE',
+            headers: {
+                'Authorization' : `Bearer ${access}`
+            },
         });
 
-        const data = await res.json();
+        const apiRes = await res.json();
 
         if (res.status === 200) {
             dispatch({
                 type: DELETE_USER_SUCCESS,
-                payload: data
+                payload: apiRes.data
             });
-            if (callback) callback([true, data.success]);
+            if (callback) callback([true, apiRes.message]);
         } else {
             dispatch({
                 type: DELETE_USER_FAIL,
-                payload: data
+                payload: apiRes.data
             });
-            if (callback) callback([false, data.error]);
+            if (callback) callback([false, apiRes.message]);
         }
     } catch (error) {
         console.log(error);
@@ -469,7 +526,7 @@ export const find_username = (email, callback) => async dispatch => {
     });
 
     try {
-        const res = await fetch('/api/user/find/username', {
+        const res = await fetch(`${API_URL}/api/user/find/username`, {
             method: 'POST',
             headers: {
                 'Accept': 'application/json',
@@ -478,25 +535,22 @@ export const find_username = (email, callback) => async dispatch => {
             body: body
         });
 
-        const data = await res.json();
+        const apiRes = await res.json();
 
         if (res.status === 200) {
-            console.log("test1");
             dispatch({
                 type: FIND_USERNAME_SUCCESS,
-                payload: data
+                payload: apiRes.data
             });
-            if (callback) callback([true, data.success]);
+            if (callback) callback([true, apiRes.message]);
         } else {
-            console.log("test2");
             dispatch({
                 type: FIND_USERNAME_FAIL,
-                payload: data
+                payload: apiRes.data
             });
-            if (callback) callback([false, data.error]);
+            if (callback) callback([false, apiRes.message]);
         }
     } catch (error) {
-        console.log("test3");
         console.log(error);
         dispatch({
             type: FIND_USERNAME_FAIL
@@ -513,7 +567,7 @@ export const reset_password = (email, new_password, new_re_password, callback) =
     });
 
     try {
-        const res = await fetch('/api/user/password/reset', {
+        const res = await fetch(`${API_URL}/api/user/password/reset`, {
             method: 'PUT',
             headers: {
                 'Accept': 'application/json',
@@ -522,20 +576,20 @@ export const reset_password = (email, new_password, new_re_password, callback) =
             body: body
         });
 
-        const data = await res.json();
+        const apiRes = await res.json();
 
         if (res.status === 200) {
             dispatch({
                 type: RESET_PASSWORD_SUCCESS,
-                payload: data
+                payload: apiRes.data
             });
-            if (callback) callback([true, data.success]);
+            if (callback) callback([true, apiRes.message]);
         } else {
             dispatch({
                 type: RESET_PASSWORD_FAIL,
-                payload: data
+                payload: apiRes.data
             });
-            if (callback) callback([false, data.error]);
+            if (callback) callback([false, apiRes.message]);
         }
     } catch (error) {
         console.log(error);
